requirements.txt
discord.py[voice]=1.2.5
